+++ src/effects/enhancers/ClassFeatures/Druid/Wildshape.mjs
@@ -6,7 +6,6 @@ export default class WildShape {
     const isLegacy = game.settings.get("dnd5e", "rulesVersion") === "legacy";
     if ((options.preset !== "wildshape") || !subject.classes?.druid) return;
     if (subject.classes.druid.subclass?.identifier === "moon" && !isLegacy) {
-      DDBEnhancers.addFeatureToEffects(subject, delta, "Circle Forms");
       DDBEnhancers.addFeatureToEffects(subject, delta, "Improved Circle Forms");
     } else if (subject.classes.druid.subclass?.identifier === "blighted") {
       DDBEnhancers.addFeatureToEffects(subject, delta, "Blighted Shape");