+++ src/parser/enrichers/class/druid/SummonWildfireSpirit.mjs
@@ -85,4 +85,5 @@ export default class SummonWildfireSpirit extends DDBEnricherData {
       },
     ];
   }
+
 }