+++ src/parser/enrichers/class/druid/WildShape.mjs
@@ -2,9 +2,8 @@
 import DDBEnricherData from "../../data/DDBEnricherData.mjs";
 
 export default class WildShape extends DDBEnricherData {
-
   get type() {
-    return "utility";
+    return "transform";
   }
 
   get activity() {
@@ -15,6 +14,50 @@ export default class WildShape extends DDBEnricherData {
           units: "hour",
         },
         img: "systems/dnd5e/icons/svg/abilities/intelligence.svg",
+        transform: {
+          customize: false,
+          identifier: "druid",
+          preset: "wildshape",
+          mode: "cr",
+        },
+        profiles: [
+          {
+            cr: "max(1/4, @subclasses.moon.levels / 3)",
+            name: "Wildshape",
+            uuid: null,
+            sizes: [],
+            types: ["beast"],
+            movement: ["fly"],
+            level: {
+              min: 2,
+              max: 3,
+            },
+          },
+          {
+            cr: "max(1/2, @subclasses.moon.levels / 3)",
+            name: "Wildshape",
+            uuid: null,
+            sizes: [],
+            types: ["beast"],
+            movement: ["fly"],
+            level: {
+              min: 4,
+              max: 7,
+            },
+          },
+          {
+            cr: "max(1, @subclasses.moon.levels / 3)",
+            name: "Wildshape",
+            uuid: null,
+            sizes: [],
+            types: ["beast"],
+            movement: [],
+            level: {
+              min: 8,
+              max: null,
+            },
+          },
+        ],
       },
     };
   }
@@ -29,9 +72,9 @@ export default class WildShape extends DDBEnricherData {
 
     if (this.is2024) {
       uses.recovery.push({
-        "period": "sr",
-        "type": "formula",
-        "formula": "1",
+        period: "sr",
+        type: "formula",
+        formula: "1",
       });
     }
 
@@ -41,5 +84,4 @@ export default class WildShape extends DDBEnricherData {
       },
     };
   }
-
 }