+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next Up
 
 - Fix issue with removed dnd5e-scriptlets function.
+- Arcane Archer Arrow corrections (2014).
+- Wild Shape and Circle Forms for 2024 Druid. (Thanks to @tminz)
 
 # 6.5.6